studiaai---learning-design-assistant (1)/backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
python-dotenv
